Added Events for the main class, a win condition function, a CharacterFactory
Game.assign_players() now uses a CharacterFactory to create the
appropriate character object. Game.get_group([]) can be used to filter
the characters in the game based on player.status, character type, or
from the latest event. When players reach max players in
game.add_player() a state change is triggered that starts the game. This
creates 2 events, day and night, from the EventFactory(), and adds them
to the list. The game.event_queue() will continue to work through the
queue until it has run out, in which case it adds another night/day, or
the game.game_end() returns a list of winners.

Much is untested. Serialisation of events to redis required, as well as
managing the Game and Player objects in memory appropriately. However,
you can now add custom events to the EventFactory, which can be
triggered through character methods.

Some minor error handling/console logging added.