Added filtering of player/user data based on who they know about
2 functions added: gain_info and lose_info.

These take action on the knows about object, which contains a list of
p_id's as keys in a dict that this player knows about. The value of the
dict is the list of attributes that the player knows. For example, if
the player is a werewolf, they would have a knows_about like so:

"1234-5678-9101112" : ["state", "character", "name"], #werewolf
"abcd-efgh-ijklmnop" : ["state", "character", "name"], #werewolf
"111..."                         : ["name"],
"112..."                         : ["name"]

When a mystic discovers the identity of a werewolf, their p_id gets
appended or altered to include the "character" attribute in the list.

Yet to interpret this info via the Game class