added filter_JSON to Game()
rudimentary filter added to Game. This takes a JSON string and replaces
the players key with a customised list hiding certain players and player
attributes. This will be cycled through and broadcast to players on
their player:p_id channels.

Quite inefficient: I serialise the whole JSON, then serialise a second
time individually for each player, with lots of duplication! :-1: